[back] test : team 생성시 teamnumber 필드 누락으로 테스트 실패 및 폰번호 수정
team_number를 만들고 추가하지 않아서 실패한 테스트들과
