+++ back/src/test/java/com/example/cpsplatform/contest/controller/service/ContestJoinServiceTest.java
@@ -65,10 +65,10 @@ class ContestJoinServiceTest {
                 .password(passwordEncoder.encode("1234"))
                 .role(Role.USER)
                 .birth(LocalDate.of(2003,1,1))
-                .email("email@email.com")
+                .email("join@email.com")
                 .address(address)
                 .gender(Gender.MAN)
-                .phoneNumber("01012341234")
+                .phoneNumber("01085292587")
                 .name("사람 이름")
                 .organization(school)
                 .build();
@@ -88,6 +88,7 @@ class ContestJoinServiceTest {
         Team team = Team.builder()
                 .name("테스트 팀")
                 .winner(false)
+                .teamNumber("001")
                 .leader(leader)
                 .contest(contest)
                 .build();
@@ -120,10 +121,10 @@ class ContestJoinServiceTest {
                 .password(passwordEncoder.encode("1234"))
                 .role(Role.USER)
                 .birth(LocalDate.of(2003,1,1))
-                .email("email@email.com")
+                .email("join@email.com")
                 .address(address)
                 .gender(Gender.MAN)
-                .phoneNumber("01012341234")
+                .phoneNumber("01074185296")
                 .name("사람 이름")
                 .organization(school)
                 .build();
@@ -162,10 +163,10 @@ class ContestJoinServiceTest {
                 .password(passwordEncoder.encode("1234"))
                 .role(Role.USER)
                 .birth(LocalDate.of(2003,1,1))
-                .email("email@email.com")
+                .email("join@email.com")
                 .address(address)
                 .gender(Gender.MAN)
-                .phoneNumber("01012341234")
+                .phoneNumber("01012348765")
                 .name("사람 이름")
                 .organization(school)
                 .build();
@@ -186,6 +187,7 @@ class ContestJoinServiceTest {
                 .name("테스트 팀")
                 .winner(false)
                 .leader(leader)
+                .teamNumber("001")
                 .contest(contest)
                 .build();
 
@@ -218,10 +220,10 @@ class ContestJoinServiceTest {
                 .password(passwordEncoder.encode("1234"))
                 .role(Role.USER)
                 .birth(LocalDate.of(2003,1,1))
-                .email("email@email.com")
+                .email("join@email.com")
                 .address(address)
                 .gender(Gender.MAN)
-                .phoneNumber("01012341234")
+                .phoneNumber("01036925814")
                 .name("사람 이름")
                 .organization(school)
                 .build();
