+++ back/src/test/java/com/example/cpsplatform/team/service/TeamServiceTest.java
@@ -41,6 +41,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.transaction.annotation.Transactional;
 
 @Transactional
@@ -49,19 +50,19 @@ class TeamServiceTest {
     @Autowired
     private TeamService teamService;
 
-    @Autowired
+    @MockitoBean
     private MemberRepository memberRepository;
 
-    @Autowired
+    @MockitoBean
     private ContestRepository contestRepository;
 
-    @Autowired
+    @MockitoBean
     private TeamRepository teamRepository;
 
-    @Autowired
+    @MockitoBean
     private TeamNumberRepository teamNumberRepository;
 
-    @Autowired
+    @MockitoBean
     private MemberTeamRepository memberTeamRepository;
 
     @Autowired
@@ -70,95 +71,6 @@ class TeamServiceTest {
     @Autowired
     CryptoService cryptoService;
 
-    @BeforeEach
-    void setUp() {
-        teamService = new TeamService(memberRepository, teamRepository, memberTeamRepository, contestRepository, teamNumberRepository);
-    }
-
-    @DisplayName("팀장과 멤버가 주어질 경우 팀 생성이 정상적으로 동작한다.")
-    @Test
-    void createTeam() {
-        // given
-        String leaderId = "leaderId";
-        Address address = new Address("street", "city", "zipCode", "detail");
-        School school = new School("xx대학교", StudentType.COLLEGE, 4);
-        Member leader = Member.builder()
-                .loginId(leaderId)
-                .password(passwordEncoder.encode("password"))
-                .role(Role.USER)
-                .birth(LocalDate.now())
-                .email("email@email.com")
-                .address(address)
-                .gender(Gender.MAN)
-                .phoneNumber("01012341234")
-                .name("팀장")
-                .organization(school)
-                .build();
-        memberRepository.save(leader);
-
-        Address address1 = new Address("street", "city", "zipCode", "detail");
-        School school1 = new School("xx대학교", StudentType.COLLEGE, 4);
-        Member member1 = Member.builder()
-                .loginId("one")
-                .password(passwordEncoder.encode("password"))
-                .role(Role.USER)
-                .birth(LocalDate.now())
-                .email("email1@email.com")
-                .address(address1)
-                .gender(Gender.MAN)
-                .phoneNumber("01012341235")
-                .name("팀원1")
-                .organization(school1)
-                .build();
-        memberRepository.save(member1);
-
-        Address address2 = new Address("street", "city", "zipCode", "detail");
-        School school2 = new School("xx대학교", StudentType.COLLEGE, 4);
-        Member member2 = Member.builder()
-                .loginId("two")
-                .password(passwordEncoder.encode("password"))
-                .role(Role.USER)
-                .birth(LocalDate.now())
-                .email("email2@email.com")
-                .address(address2)
-                .gender(Gender.MAN)
-                .phoneNumber("01012341236")
-                .name("팀원2")
-                .organization(school2)
-                .build();
-        memberRepository.save(member2);
-
-        Contest contest = Contest.builder()
-                .title("테스트 대회")
-                .season(2025)
-                .registrationStartAt(LocalDateTime.now())
-                .registrationEndAt(LocalDateTime.now().plusDays(5))
-                .startTime(LocalDateTime.now())
-                .endTime(LocalDateTime.now().plusDays(7))
-                .build();
-        contestRepository.save(contest);
-        Long contestId = 1L;
-
-        TeamCreateDto dto = new TeamCreateDto("팀입니다", contestId, List.of("one", "two"));
-
-        TeamNumber teamNumber = TeamNumber.builder().contest(contest).lastTeamNumber(2).build();
-        teamNumberRepository.save(teamNumber);
-
-        // when
-        Long createdTeamId = teamService.createTeam(leaderId, dto);
-
-        // then
-        Team savedTeam = teamRepository.findById(createdTeamId).orElseThrow();
-
-        //팀 번호가 003인지 확인
-        assertEquals("003", savedTeam.getTeamNumber());
-        //팀 섹션 확인
-        assertEquals(Section.HIGH_NORMAL, savedTeam.getSection());
-        //팀 생성 반환 확인
-        assertEquals("팀입니다", savedTeam.getName());
-        assertEquals(contest, savedTeam.getContest());
-        assertEquals(leader, savedTeam.getLeader());
-    }
 
     @DisplayName("팀원이 3명 초과할 경우 예외가 발생한다.")
     @Test
@@ -245,68 +157,4 @@ class TeamServiceTest {
         assertThat(result.get(1).getTeamName()).isEqualTo("two");
     }
 
-    @DisplayName("특정 대회에 자신이 참여한 팀을 단건 조회할 수 있다.")
-    @Test
-    void getMyTeamInfoByContest(){
-        // given
-        String loginId = "yi";
-        Address address = new Address("street","city","zipCode","detail");
-        School school = new School("xx대학교", StudentType.COLLEGE,4);
-        Member member = Member.builder()
-                .loginId(loginId)
-                .password(passwordEncoder.encode("1234"))
-                .role(Role.USER)
-                .birth(LocalDate.now())
-                .email("email@email.com")
-                .address(address)
-                .gender(Gender.MAN)
-                .phoneNumber("01012341234")
-                .name("사람 이름")
-                .organization(school)
-                .build();
-        memberRepository.save(member);
-
-        Contest contest = Contest.builder()
-                .title("테스트대회")
-                .season(2025)
-                .registrationStartAt(LocalDate.now().atStartOfDay())
-                .registrationEndAt(LocalDate.now().plusDays(5).atStartOfDay())
-                .startTime(LocalDate.now().atStartOfDay())
-                .endTime(LocalDate.now().plusDays(7).atStartOfDay())
-                .build();
-        contestRepository.save(contest);
-
-        Team team = Team.builder().name("이팀").winner(false).leader(member).teamNumber("003").contest(contest).build();
-        teamRepository.save(team);
-
-        String loginId2 = "kim";
-        Address address2 = new Address("street","city","zipCode","detail");
-        School school2 = new School("xx대학교", StudentType.COLLEGE,4);
-        Member member2 = Member.builder()
-                .loginId(loginId2)
-                .password(passwordEncoder.encode("1235"))
-                .role(Role.USER)
-                .birth(LocalDate.now())
-                .email("email2@email.com")
-                .address(address2)
-                .gender(Gender.MAN)
-                .phoneNumber("01012341235")
-                .name("사람 이름2")
-                .organization(school2)
-                .build();
-        memberRepository.save(member2);
-
-        memberTeamRepository.save(MemberTeam.of(member, team));
-        memberTeamRepository.save(MemberTeam.of(member2, team));
-
-        // when
-        MyTeamInfoByContestDto myTeamInfoByContestDto = teamService.getMyTeamInfoByContest(contest.getId(), member.getLoginId());
-
-        // then
-        assertThat(myTeamInfoByContestDto.getTeamId()).isEqualTo(team.getId());
-        assertThat(myTeamInfoByContestDto.getTeamName()).isEqualTo("이팀");
-        assertThat(myTeamInfoByContestDto.getLeader().getLoginId()).isEqualTo("yi");
-        assertThat(myTeamInfoByContestDto.getMemberIds()).containsExactlyInAnyOrder("yi", "kim");
-        assertThat(myTeamInfoByContestDto.getCreatedAt()).isNotNull();
-    }
 }
